style: Modified class row to have margin on left and right of 5%
Modified the bootstrap row margin from -15px to 5% since it gives it a better margin on all sizes when
the screens are resized. Also, deleted the max-width of body when screen is bigger then 992px because I only
had it at max-width of 992px when screen was bigger then 992px.