style: Modified the main picture when resizing screen
The main pic on website was too much to the left so went ahead and
changed the css to center a bit more the main pic. The picture was too mib and since it was more to
the left it gave users a horizontal scrollbar but there wasnt much on left side just extra space from the leftover of
mainpic.