style: Modified the opacity of the main Pic
Decided to change the opacity of main pic b/c it will have users eyes focus on
the other elements which are the pics of the projects and their respective URLs.

This shows true for the smaller screens since the main pic is the biggest picture in the web site and the first. When users
now come to visit the website it will focus on the name and rest of the site instead of just focusing on the main pic.