Feat: Added new tab in navbar
Added new tab in navbar. For now this is a place holder.
But in future will be adding it as a dropdown with list of all project I have completed. Where if project has an
live url it will take them there if not it will sent them to github repo.